Update scaling_rag_for_production.md
edited with input from Pascal:

line 391 - “This simplified version is adapted to our code and leaves out a bunch of advanced retrieval techniques, such as reranking and hybrid search. We use gpt-3.5-turbo as our LLM and query it via the OpenAI API.”
Q1: what does “our code” refer to specifically? Do you just mean that we have adapted generate.py to simplify it? Or is something more implied by “adapted to our code”?
A1: Just small changes in order to make it work with the rest of the script, such as naming and also some minor simplifications to make it easier to read/understand.
Q2: why do we leave out a bunch of advanced retrieval techniques?
A2: Because they weren't relevant to the topic of how to scale a RAG system. There are a lot of different techniques that one can use that might or might not add to performance. I felt that it wouldn't have added value in the context of the article to introduce "unnecessary" elements.
line 470 -
“even more convenient” - Q: how, specifically? how does implementation within a single QueryAgent class do this?
A: Because with a single class, we have everything we need bundled together and there's no need for sequentially calling functions (if I understood the question correctly).
“take care of all the steps we implemented above” - Q: specically, which steps are you referring to?
A: Everything we laid out before: getting the embeddings, embedding the search query, performing vector search, processing the results and querying the LLM to generate a response.
“including a few additional utility functions” - Q: I see the get_num_tokens(text) and trim(text, max_context_length) utility functions… can you indicate briefly what specifically they do in our generation step here?
A: Get_num_tokens encodes a text and gets the number of tokens. This is done to calculate the length of the input. We don't want the input to be too long, so that there's still room for the generations. We use standard values of 50/50 - up to 50% of the total context for the input and up to 50% for the generations. Trim simply trims the input text if it's too long to fit into the context.
line 552 -
“And this is how we can use the QueryAgent:”
Q: you mean to generate a response? at this point, we have already contextualized the query, correct? so now we are generating a response to a contextualized query?
A: QueryAgent includes all of the previous steps too. It's just a more convenient way to run the whole thing. So yes, to generate a response, but also to embed our query and retrieve relevant vectors (is this what you mean with contextualized?) before we generate the response with those.
finally, in the indexing section:
using Qdrant’s new in-memory feature.. can you say just a little more about using this instead of a tool suited to productionizing…
i.e., if the tutorial is supposed to show a way of developing to productionize all at once, this seems like a weakness. If we can point to a solution that might work in a production environment, at least we have that covered.
A: Everything would be the same, we'd just have to point to our Qdrant instance instead of using it in-memory. Maybe we can add a sentence to clarify this? A lot of people are using hosted vendor instances, that would mean they would simply take their URL and enter it there. Self-hosting a Kubernetes cluster or similar local deployments database would be outside of the tutorial I think.





12:45
PS: Hope that's readable enough, I answered below the questions


robertdhayanturner
  1:57 PM
okay, great. I’ve gone thru and edited everything accordingly. I’ll link you on my next commit. Pls as soon as you can (hopefully today), check the final edit and lmk if anything needs changing before publishing.
1:57
https://github.com/superlinked/VectorHub/blob/main/docs/use_cases/scaling_rag_for_production.md


robertdhayanturner
  2:04 PM
in the meantime, I’m going thru again to catch any grammatical / diction errors; don’t worry about those - I’ll get them. I just need your input on content.
I’m also removing a couple notes to myself I forgot about while editing - ignore those. (edited) 


Pascal Biese
  2:31 PM
Thanks! Only thing I would change are some minor wordings I used in this chat that weren't refined enough to use in an article. In line 293, I would change "self-hosting a Kubernetes cluster or similar local deployments database" to something like "[Detailed guidance on] self-hosted solutions, such as setting up a self-hosted Kubernetes cluster, is beyond the scope of this tutorial". In line 391, I would change "to simplify and keep our focus on scalable RAG" to "to simplify and keep our focus on how to scale a basic RAG system"
Everything else looks fine!